LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/anchor-transition-001-expected.txt


FAIL Transition of anchor() when changing target anchor element name assert_equals: expected 300 but got 225
FAIL Transition of anchor-size() when changing target anchor element name assert_equals: expected 150 but got 100
